Work around BMC #18209
A pointless windowiconprovider lives in meego-ux-components that is listening
for a signal on qApp for updateWindowList.  The code (which we don't actually use)
that is triggered by this signal has a horrific memory leak, so after a few window
list updates the system is chocking for free memory.

To work around this we just call the our own method for operating on the window
list instead of catching our own signal.