Fix BMC #18763
The music app is installing a dbus autostart configuration, which
for some reason results in the app immediately exiting after starting
from the launcher, and then starting again.  The result exposed an
unexpected corner case for the logic that attempt so map unknown windows
to desktop files.

The fix was to set the desktop pid in addition to it's window id when
a fuzzy name match is made.